requirements.txt
boto3~=1.28.75
pandas~=1.2.5
botocore~=1.31.75
requests~=2.26.0
openpyxl~=3.0.7
gitpython
sqlalchemy
pymysql